Sensory Integration and ADHD Educational Recommendation System/requirements.txt
streamlit
plotly
pandas
numpy
scikit-learn
typing-extensions
PyPDF2
python-docx 
nltk 
spacy 
scikit-learn 